refactor(solana): move transaction signing out of :libs:solana:encoding
SolanaTransaction.sign/signatures depended on com.getcode.ed25519.Ed25519,
the Android-only JNI wrapper (its KeyPair is Parcelable), not the KMP
Ed25519Kmp. That import kept :libs:solana:encoding from compiling on the
Apple targets, since the module is meant to be encoding only.

Move both functions to services/opencode/src/main/kotlin/com/getcode/opencode/solana/SolanaTransactionSigning.kt
as extensions on SolanaTransaction, unchanged, alongside SigningError. Ed25519
stays the legacy JNI class; nothing migrates to Ed25519Kmp here.
