+++ services/opencode/src/main/kotlin/com/getcode/opencode/internal/network/api/intents/IntentStatefulSwap.kt
@@ -16,6 +16,7 @@ import com.getcode.opencode.model.transactions.SwapProgram
 import com.getcode.opencode.model.transactions.VerifiedSwapMetadata
 import com.getcode.opencode.solana.SolanaTransaction
 import com.getcode.opencode.solana.TransactionBuilder
+import com.getcode.opencode.solana.signatures
 import com.getcode.solana.keys.Signature
 
 internal class IntentStatefulSwap(
